requirements.txt
ollama>=0.3.0
playwright>=1.44.0
python-docx>=1.1.0
streamlit>=1.35.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pandas>=2.2.0
lxml>=5.1.0
numpy==1.26.4
langgraph>=0.2.0
langchain-core>=0.3.0